requirements.txt
Flask==2.3.3
gunicorn>=21.0.0
Pillow>=10.2.0
Werkzeug==2.3.7
Flask-CORS==4.0.0
python-dotenv==1.0.0
pydub==0.25.1
# Python 3.13+ 已移除 stdlib audioop，pydub 需此兼容包（提供 audioop）
audioop-lts>=0.2.0; python_version>="3.13"
# openai 2.x 需较新 typing_extensions（过旧会导致无法从 typing_extensions 导入 List 等）
typing_extensions>=4.12.0
# httpx>=0.28 移除 proxies 参数，旧版 openai 会报错；与 openai 1.56+ 或限制 httpx 二选一即可
openai>=1.56.0
httpx>=0.27.0,<0.28.0
requests>=2.28.0
# 腾讯新闻频道页懒加载 & 图片抓取：playwright 无头渲染（安装后还需 playwright install chromium）
playwright>=1.45.0
python-docx>=1.1.0
# Word 文档文稿上传补全（.docx）依赖 Aspose.Words（商业许可需自行配置）
aspose-words>=24.0.0
openpyxl>=3.1.0
reportlab>=4.0.0
# Knowledge base (RAG) vector store
chromadb>=0.5.0
# LabelImg 数据标注（桌面 GUI 工具）
labelImg
PyQt5
lxml
ultralytics
torch
onnx
tensorflow>=2.16
keras>=3.0
